refactor: 오늘의 기도 컴포넌트 분리
